feat(actions): generate and embed the actions documentation
