+++ internal/cmd/.gitignore
@@ -1,0 +1,1 @@
+actions-help.txt
